fix property types

+++ src/ForwardingController/CrossCutting/ForwardingUtils.SharedItems/JsonModel/Entity/HttpEntity.cs
@@ -2,14 +2,14 @@
 {
     public class HttpEntity
     {
-        public int count { get; set; }
-        public int rate1 { get; set; }
-        public int rate5 { get; set; }
-        public int rate15 { get; set; }
-        public int p50 { get; set; }
-        public int p90 { get; set; }
-        public int p95 { get; set; }
-        public int p99 { get; set; }
+        public decimal count { get; set; }
+        public decimal rate1 { get; set; }
+        public decimal rate5 { get; set; }
+        public decimal rate15 { get; set; }
+        public decimal p50 { get; set; }
+        public decimal p90 { get; set; }
+        public decimal p95 { get; set; }
+        public decimal p99 { get; set; }
 
         public HttpEntity()
         {
